update 2021/11/10 incomplete
todo:
remove user (uid) from most request struct
use that from token instead
change frontend default view to make it fetch token in the beginning

others:
create new users (profile + indexes of users update)(may be add a function called save indexesOfUsersToDisk)

redo and undo

+++ simple-webpage-document-storage-sys/controller/post.go
@@ -3,20 +3,77 @@ package controller
 import (
 	"errors"
 	"github.com/gin-gonic/gin"
-	"math/rand"
 	"net/http"
 	"simple-webpage-document-storage-sys/common"
 	"simple-webpage-document-storage-sys/logging"
 	"simple-webpage-document-storage-sys/manager"
+	"simple-webpage-document-storage-sys/token"
 	"strconv"
-	"time"
 )
+
+// verify checks if the uid in the token matches the uid in the request
+func verify(c *gin.Context, userId string) bool {
+	uid, exist := c.Get(common.TokenUid)
+	if (!exist && userId != common.DefaultUserId) || uid != userId {
+		return false
+	}
+	return true
+}
+
+// Login for logging in
+func Login(c *gin.Context) {
+	request := &RequestLogin{}
+	err := c.BindJSON(request)
+	logging.ConditionalLogError(err)
+
+	uid, b := manager.VerifyUserPassword(request.Username, request.Password)
+
+	if !b {
+		c.JSON(http.StatusOK, &LoginResponse{Success: false, User: "", Token: ""})
+		return
+	}
+
+	tk, err := token.GenerateToken(uid)
+	if err != nil {
+		logging.Error(err)
+		return
+	}
+
+	manager.RegisterUser(uid)
+
+	c.JSON(http.StatusOK, &LoginResponse{Success: true, User: uid, Token: tk})
+}
+
+// Logout for logging out
+func Logout(c *gin.Context) {
+	request := &RequestLogout{}
+	err := c.BindJSON(request)
+	logging.ConditionalLogError(err)
+
+	uid, b := token.DecodeToken(request.Token)
+
+	if !b || uid != request.User {
+		logging.Error(errors.New(errorInvalidToken),
+			logging.S("uid", request.User), logging.S("uid from token", uid), logging.S("token", request.Token))
+		c.Status(http.StatusBadRequest)
+		return
+	}
+
+	manager.UnregisterUser(uid)
+	c.Status(http.StatusOK)
+}
 
 // GetFile deals with requests that ask for a specific file
 func GetFile(c *gin.Context) {
 	request := &RequestFile{}
 	err := c.BindJSON(request)
 	logging.ConditionalLogError(err)
+
+	if !verify(c, request.User) {
+		c.Status(http.StatusBadRequest)
+		return
+	}
+
 	fileName, content := manager.FetchTxt(request.User, request.Fid)
 	c.JSON(http.StatusOK, &FileResponse{FileName: fileName, Content: content})
 }
@@ -26,6 +83,12 @@ func ModifyFile(c *gin.Context) {
 	request := &RequestModifyFile{}
 	err := c.BindJSON(request)
 	logging.ConditionalLogError(err)
+
+	if ! verify(c, request.User) {
+		c.Status(http.StatusBadRequest)
+		return
+	}
+
 	b := manager.ModifyTxt(request.User, request.Fid, request.NewC)
 	if b {
 		c.Status(http.StatusOK)
@@ -42,6 +105,12 @@ func Rename(c *gin.Context) {
 	request := &RequestRename{}
 	err := c.BindJSON(request)
 	logging.ConditionalLogError(err)
+
+	if !verify(c, request.User) {
+		c.Status(http.StatusBadRequest)
+		return
+	}
+
 	b := false
 	if request.Dir {
 		b = manager.RenameDir(request.User, request.ObjId, request.NewName)
@@ -58,23 +127,19 @@ func Rename(c *gin.Context) {
 	}
 }
 
-func generateRandomId(n int) string {
-	id := make([]byte, n)
-	length := len(common.LetterBytes)
-	rand.Seed(time.Now().UnixNano())  // to avoid same random ids
-	for i := range id {
-		id[i] = common.LetterBytes[rand.Intn(length)]
-	}
-	return string(id)
-}
-
 // Create deals with requests that create a new file or a new dir
 func Create(c *gin.Context) {
 	request := &RequestCreate{}
 	err := c.BindJSON(request)
 	logging.ConditionalLogError(err)
+
+	if !verify(c, request.User) {
+		c.Status(http.StatusBadRequest)
+		return
+	}
+
 	b := false
-	objId := generateRandomId(64)
+	objId := common.GenerateRandomId(64)
 	if request.Dir {
 		b = manager.CreateDir(request.User, objId, request.Name, request.ParentId)
 	} else {
@@ -96,6 +161,12 @@ func Delete(c * gin.Context) {
 	request := &RequestDelete{}
 	err := c.BindJSON(request)
 	logging.ConditionalLogError(err)
+
+	if !verify(c, request.User) {
+		c.Status(http.StatusBadRequest)
+		return
+	}
+
 	b := false
 	if request.Dir {
 		b = manager.DeleteDir(request.User, request.ObjId)
@@ -117,6 +188,12 @@ func Move(c *gin.Context) {
 	request := &RequestMove{}
 	err := c.BindJSON(request)
 	logging.ConditionalLogError(err)
+
+	if !verify(c, request.User) {
+		c.Status(http.StatusBadRequest)
+		return
+	}
+
 	b := false
 	if request.Dir {
 		b = manager.MoveDir(request.User, request.ObjId, request.NewParentId)
